chore(jsdoc): add jsdocs for dashboard component
